move 'View profile of all members in a group' to other file
